+++ temba/flows/legacy/tests.py
@@ -130,17 +130,20 @@ class FlowMigrationTest(TembaTest):
         flow_json = self.load_flow_def("malformed_single_message")["definition"]
 
         FlowRevision.objects.create(flow=flow, definition=flow_json, spec_version=3, revision=1, created_by=self.admin)
+        revision = flow.revisions.get()
 
-        flow.ensure_current_version()
-        flow_json = flow.get_definition()
+        revision.get_migrated_definition()
 
-        self.assertEqual(1, len(flow_json["nodes"]))
-        self.assertEqual(Flow.CURRENT_SPEC_VERSION, flow_json["spec_version"])
-        self.assertEqual(2, flow_json["revision"])
+        # rapidpro should have wrapped the pre-v6 revision, run the legacy migration chain, and handed the result
+        # off to mailroom to migrate up to the current spec (the migration itself is goflow's job)
+        self.assertEqual(1, len(mr_mocks.calls["flow_migrate"]))
+        handed_off, to_version = mr_mocks.calls["flow_migrate"][0].args
+        self.assertEqual(Flow.FINAL_LEGACY_VERSION, handed_off["version"])
+        self.assertEqual(Flow.CURRENT_SPEC_VERSION, to_version)
 
     @mock_mailroom
     def test_migrate_to_11_12(self, mr_mocks):
-        flow = self.load_flow("favorites")
+        flow = self.create_flow("Favorites")
         definition = {
             "entry": "79b4776b-a995-475d-ae06-1cab9af8a28e",
             "rule_sets": [],
@@ -233,7 +236,7 @@ class FlowMigrationTest(TembaTest):
         # removed the invalid reference
         self.assertEqual(len(migrated["action_sets"]), 2)
 
-        flow = self.load_flow("migrate_to_11_12")
+        flow = self.create_flow("Migrate 11.12")
         flow_json = self.load_flow_def("migrate_to_11_12")
         migrated = migrate_to_version_11_12(flow_json, flow)
 
@@ -242,7 +245,7 @@ class FlowMigrationTest(TembaTest):
 
     @mock_mailroom
     def test_migrate_to_11_12_with_one_node(self, mr_mocks):
-        flow = self.load_flow("migrate_to_11_12_one_node")
+        flow = self.create_flow("Migrate 11.12 One Node")
         flow_json = self.load_flow_def("migrate_to_11_12_one_node")
         migrated = migrate_to_version_11_12(flow_json, flow)
 
@@ -250,7 +253,7 @@ class FlowMigrationTest(TembaTest):
 
     @mock_mailroom
     def test_migrate_to_11_12_other_org_existing_flow(self, mr_mocks):
-        flow = self.load_flow("migrate_to_11_12_other_org", {"CHANNEL-UUID": str(self.channel.uuid)})
+        flow = self.create_flow("Migrate 11.12 Other Org")
         flow_json = self.load_flow_def("migrate_to_11_12_other_org", {"CHANNEL-UUID": str(self.channel.uuid)})
 
         # change ownership of the channel it's referencing
@@ -262,18 +265,9 @@ class FlowMigrationTest(TembaTest):
         # check action set was removed
         self.assertEqual(len(migrated["rule_sets"]), 0)
 
-    def test_migrate_to_11_12_channel_dependencies(self):
-        self.channel.name = "1234"
-        self.channel.save()
-
-        self.load_flow("migrate_to_11_12_one_node")
-        flow = Flow.objects.filter(name="channel").first()
-
-        self.assertEqual(flow.channel_dependencies.count(), 1)
-
     @mock_mailroom
     def test_migrate_to_11_11(self, mr_mocks):
-        flow = self.load_flow("migrate_to_11_11")
+        flow = self.create_flow("Migrate 11.11")
         flow_json = self.load_flow_def("migrate_to_11_11")
 
         migrated = migrate_to_version_11_11(flow_json, flow)
@@ -334,11 +328,15 @@ class FlowMigrationTest(TembaTest):
 
     @mock_mailroom
     def test_migrate_to_11_9(self, mr_mocks):
-        flow = self.load_flow("migrate_to_11_9", name="Master")
+        flow = self.create_flow("Master")
 
-        # give our flows same UUIDs as in import and make 2 of them invalid
+        # the migration drops references to flows that are missing or invalid - create the referenced flows with
+        # the UUIDs the definition uses and make 2 of them invalid
+        self.create_flow("Valid1")
         Flow.objects.filter(name="Valid1").update(uuid="b823cc3b-aaa6-4cd1-b7a5-28d6b492cfa3")
+        self.create_flow("Invalid1")
         Flow.objects.filter(name="Invalid1").update(uuid="ad40071e-a665-4df3-af14-0bc0fe589244", is_archived=True)
+        self.create_flow("Invalid2")
         Flow.objects.filter(name="Invalid2").update(uuid="136cdab3-e9d1-458c-b6eb-766afd92b478", is_active=False)
 
         import_def = self.load_json("test_flows/legacy/migrations/migrate_to_11_9.json")
@@ -386,8 +384,12 @@ class FlowMigrationTest(TembaTest):
         self.assertEqual(len(migrated["rule_sets"]), 6)
 
     def test_migrate_to_11_6(self):
-        flow = self.load_flow("migrate_to_11_6")
+        flow = self.create_flow("Migrate 11.6")
         flow_json = self.load_flow_def("migrate_to_11_6")
+
+        # the migration remaps the definition's group references to org groups looked up by name
+        for name in set(get_legacy_groups(flow_json).values()):
+            self.create_group(name, contacts=[])
 
         migrated = migrate_to_version_11_6(flow_json, flow)
         migrated_groups = get_legacy_groups(migrated)
@@ -656,7 +658,7 @@ class FlowMigrationTest(TembaTest):
         self.create_field("district", "District", ContactField.TYPE_DISTRICT)
         self.create_field("joined_on", "Joined On", ContactField.TYPE_DATETIME)
 
-        flow = self.load_flow("type_flow")
+        flow = self.create_flow("Type Flow")
         flow_def = self.load_flow_def("type_flow")
         migrated = migrate_to_version_11_0(flow_def, flow)
 
@@ -686,7 +688,7 @@ class FlowMigrationTest(TembaTest):
 
     @mock_mailroom
     def test_migrate_to_11_0_with_null_ruleset_label(self, mr_mocks):
-        flow = self.load_flow("migrate_to_11_0")
+        flow = self.create_flow("Migrate 11.0")
         definition = {
             "rule_sets": [
                 {
@@ -708,7 +710,7 @@ class FlowMigrationTest(TembaTest):
 
     @mock_mailroom
     def test_migrate_to_11_0_with_null_msg_text(self, mr_mocks):
-        flow = self.load_flow("migrate_to_11_0")
+        flow = self.create_flow("Migrate 11.0")
         definition = {
             "action_sets": [
                 {
@@ -726,7 +728,7 @@ class FlowMigrationTest(TembaTest):
 
     @mock_mailroom
     def test_migrate_to_11_0_with_broken_localization(self, mr_mocks):
-        flow = self.load_flow("migrate_to_11_0")
+        flow = self.create_flow("Migrate 11.0")
         flow_def = self.load_flow_def("migrate_to_11_0")
         migrated = migrate_to_version_11_0(flow_def, flow)
 
@@ -782,7 +784,7 @@ class FlowMigrationTest(TembaTest):
     @mock_mailroom
     def test_migrate_to_10(self, mr_mocks):
         # this is really just testing our rewriting of webhook rulesets
-        flow = self.load_flow("dual_webhook")
+        flow = self.create_flow("Dual Webhook")
         flow_def = self.load_flow_def("dual_webhook")
 
         # get our definition out
@@ -1054,27 +1056,6 @@ class FlowMigrationTest(TembaTest):
         email_action = email_node["actions"][1]
 
         self.assertEqual(["admin@textit.com"], email_action["addresses"])
-
-    def test_migrate_bad_group_names(self):
-        # This test makes sure that bad contact groups (< 25, etc) are migrated forward properly.
-        # However, since it was a missed migration, now we need to apply it for any current version
-        # at the time of this fix
-        for v in ("4", "5", "6", "7", "8", "9", "10"):
-            error = 'Failure migrating group names "%s" forward from v%s'
-            flow = self.load_flow("favorites_bad_group_name_v%s" % v)
-            self.assertIsNotNone(flow, "Failure importing favorites from v%s" % v)
-            self.assertTrue(ContactGroup.objects.filter(name="Contacts < 25").exists(), error % ("< 25", v))
-            self.assertTrue(ContactGroup.objects.filter(name="Contacts > 100").exists(), error % ("> 100", v))
-
-            ContactGroup.objects.filter(is_system=False).delete()
-            self.assertEqual(Flow.CURRENT_SPEC_VERSION, flow.version_number)
-            flow.release(self.admin, interrupt_sessions=False)
-
-    def test_migrate_malformed_groups(self):
-        flow = self.load_flow("malformed_groups")
-        self.assertIsNotNone(flow)
-        self.assertTrue(ContactGroup.objects.filter(name="Contacts < 25").exists())
-        self.assertTrue(ContactGroup.objects.filter(name="Unknown").exists())
 
 
 class MigrationUtilsTest(TembaTest):